Readme left on my Google drive regarding this version: This version has a setup that allows for drawing lists of separate emitters in one go. It's done by identifying the emitters by their generated ID and then drawing all the ParticleData with that ID in one go. It's nice and easy to work with but suffers in performance compared to other methods of drawing particles on separate RenderTargets.
It has to basically loop over all the particles in the RenderManager only drawing specific particles and then loop over them all over again for the next list of emitters. So in this version it loops over all particles 3 times to check IDs before the frame is finished.

The problem is then further compounded by having to repeat that process for every emitter in the list so if there are 40 emitters, the loop over every particle actually happens 120 times. There is no update or draw code actually executed, but it still has to loop over each particle using the List.Where to check if the particles match the ID

The result is only being able to render approx. 4000 particles before losing frames as opposed to 20,000+ particles.